+++ tests/unit/shutdown.tcl
@@ -25,33 +25,35 @@ start_server {tags {"shutdown"}} {
     }
 }
 
-start_server {tags {"shutdown"}} {
-    test {Temp rdb will be deleted in signal handle} {
-        for {set i 0} {$i < 20} {incr i} {
-            r set $i $i
-        }
-        # It will cost 2s (20 * 100ms) to dump rdb
-        r config set rdb-key-save-delay 100000
-        
-        set pid [s process_id]
-        set temp_rdb [file join [lindex [r config get dir] 1] temp-${pid}.rdb]
+if {$::tcl_platform(platform) ne "windows"} {
+    start_server {tags {"shutdown"}} {
+        test {Temp rdb will be deleted in signal handle} {
+            for {set i 0} {$i < 20} {incr i} {
+                r set $i $i
+            }
+            # It will cost 2s (20 * 100ms) to dump rdb
+            r config set rdb-key-save-delay 100000
 
-        # trigger a shutdown which will save an rdb
-        exec kill -SIGINT $pid
-        # Wait for creation of temp rdb
-        wait_for_condition 50 10 {
-            [file exists $temp_rdb]
-        } else {
-            fail "Can't trigger rdb save on shutdown"
-        }
+            set pid [s process_id]
+            set temp_rdb [file join [lindex [r config get dir] 1] temp-${pid}.rdb]
 
-        # Insist on immediate shutdown, temp rdb file must be deleted
-        exec kill -SIGINT $pid
-        # wait for the rdb file to be deleted
-        wait_for_condition 50 10 {
-            ![file exists $temp_rdb]
-        } else {
-            fail "Can't trigger rdb save on shutdown"
+            # trigger a shutdown which will save an rdb
+            exec kill -SIGINT $pid
+            # Wait for creation of temp rdb
+            wait_for_condition 50 10 {
+                [file exists $temp_rdb]
+            } else {
+                fail "Can't trigger rdb save on shutdown"
+            }
+
+            # Insist on immediate shutdown, temp rdb file must be deleted
+            exec kill -SIGINT $pid
+            # wait for the rdb file to be deleted
+            wait_for_condition 50 10 {
+                ![file exists $temp_rdb]
+            } else {
+                fail "Can't trigger rdb save on shutdown"
+            }
         }
     }
 }
